Add files via upload
A set of automatic dew point gas generation system is designed based on S7-200 smart PLC. The software of the project completes the design of data acquisition, processing, signal control and transmission, and upper computer configuration. According to the user's requirements, combined with the ambient temperature, pressure and target dew point, the corresponding dry and wet theoretical flow rate is calculated and the 0-5V voltage signal is output to drive the corresponding mass flow controller to realize the occurrence of the target dew point.